move file preview logic to dedicated service
