+++ libs/shared/src/lib/components/file-preview-modal/file-preview-modal.component.scss
